src/requirements.txt
pyzmq
pymongo
pyyaml